article_microservices/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
newsapi==0.1.1
newsapi-python==0.2.6
pandas==1.4.0
numpy==1.22.0
requests==2.27.1
gremlinpython==3.4.10
bs4
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.13.1+cpu
sentence-transformers
transformers
asyncio
nest-asyncio